ESP Configuration
This was ESP32-Cam configuration. You can use this or using example on Arduino IDE. The purpose of this ESP32-Cam configuration is just to stream the camera on local host. You can use any Router access point without connected to internet. The camera will stream trough local host and accessed on laptop or desktop on the same wifi and use it to get the video streams.